Extract stats player config runtime helpers
